moved free_slots endpoint to psycho modul
